IGNITE-27944 Fix broken ZookeeperDiscoverySpiSslTest.
